Move FieldKit from botany to genetics module
