+++ src/main/resources/assets/botany/lang/en_US.lang
@@ -158,10 +158,6 @@ botany.item.tooltip.fertility=Fertility: %dx
 botany.item.tooltip.flowerColor1=%1$s, %2$s Stem
 botany.item.tooltip.flowerColor2=%1$s and %2$s, %3$s Stem
 
-botany.item.fieldKit.tooltip.noPaper=No paper
-botany.item.fieldKit.tooltip.sheetsOfPaper=%d sheets of paper
-botany.item.fieldKit.tooltip.sheetOfPaper=%d sheet of paper
-
 botany.item.soilMeter.name=Soil Meter
 botany.item.pigment.name=Pigment
 botany.item.clay.name=Clay
@@ -171,7 +167,6 @@ botany.item.germling.tag.name=Germling
 botany.item.flowerCorrupted.name=Corrupted Flower
 botany.item.database.0.name=Botanist Database
 botany.item.database.1.name=Master Botanist Database
-botany.item.fieldKit.name=Field Kit
 botany.item.powderAsh.name=Ash Powder
 botany.item.powderPulp.name=Pulp Powder
 botany.item.powderMulch.name=Mulch Powder
