+++ src/main/resources/assets/genetics/lang/zh_CN.lang
@@ -4,6 +4,11 @@ fluid.binnie.growthMedium=培养基
 fluid.binnie.bacteriaPoly=聚合菌
 fluid.binnie.bacteriaVector=细菌质粒
 
+genetics.item.fieldKit.name=实地考察工具包
+genetics.item.fieldKit.tooltip.noPaper=没有纸
+genetics.item.fieldKit.tooltip.sheetsOfPaper=%d 张纸
+genetics.item.fieldKit.tooltip.sheetOfPaper=%d 张纸
+
 genetics.item.casingIron.name=强化外壳
 genetics.item.dnaDye.name=DNA 标记染料
 genetics.item.dyeFluor.name=荧光染料
